requirements.txt
pandas
requests
python-dotenv
numpy
matplotlib
seaborn
scikit-learn
